Unify the handling of "X as combinator" - use an array of int, rather than one global int per combinator.  This allows another change, the unification of communication between lexer and parser about file names, for the "load" interpreter command.  Now, unless it has escape sequences, or whitespace, a filename does not have to have quotes.
Unify the handling of interpreter commands that used to toggle output
(trace, debug, etc), but now have an "on" or "off" modifier.  Those same
commands print the current state, when given without on or off modifiers.
As a further goodie, "on" and "off" can also appear as file names, after
a "load" command.